+++ SharePointDsc/Examples/Resources/SPManagedMetaDataServiceAppDefault/1-SetManagedMetadataDefault.ps1
@@ -35,7 +35,8 @@ Updated author, copyright notice, and URLs.
 <#
 
 .DESCRIPTION
- This example shows how to deploy the Managed Metadata service app to the local SharePoint farm.
+ This example shows how to configure a default Managed Metadata service app for the default
+ proxy group.
 
 #>
 
@@ -51,7 +52,7 @@ Configuration Example
     node localhost {
         SPManagedMetaDataServiceAppDefault ManagedMetadataServiceAppDefault
         {
-            IsSingleInstance               = "Yes"
+            ServiceAppProxyGroup           = "Default"
             DefaultSiteCollectionProxyName = "Managed Metadata Service Application Proxy"
             DefaultKeywordProxyName        = "Managed Metadata Service Application Proxy"
             PsDscRunAsCredential           = $SetupAccount